Rename the VM C files according to the Great Rocqification.
