+++ kernel/byterun/rocq_memory.h
