+++ kernel/genOpcodeFiles.ml
@@ -10,12 +10,12 @@
 
 (** List of opcodes.
 
-    It is used to generate the files [coq_instruct.h], [coq_jumptbl.h],
-    [coq_arity.h], and [vmopcodes.ml].
+    It is used to generate the files [rocq_instruct.h], [rocq_jumptbl.h],
+    [rocq_arity.h], and [vmopcodes.ml].
 
     [STOP] needs to be the last opcode.
 
-    Arity -1 designates opcodes that need special handling in [coq_fix_code.c].
+    Arity -1 designates opcodes that need special handling in [rocq_fix_code.c].
 *)
 let opcodes =
   [|
@@ -173,7 +173,7 @@ let pp_header isOcaml fmt =
     (if isOcaml then pp_ocaml_comment else pp_c_comment)
     "DO NOT EDIT: automatically generated by kernel/genOpcodeFiles.ml"
 
-let pp_coq_instruct_h fmt =
+let pp_rocq_instruct_h fmt =
   pp_header false fmt;
   Format.fprintf fmt "#pragma once@.enum instructions {@.";
   Array.iter (fun (name, _) ->
@@ -181,13 +181,13 @@ let pp_coq_instruct_h fmt =
     ) opcodes;
   Format.fprintf fmt "};@."
 
-let pp_coq_jumptbl_h fmt =
+let pp_rocq_jumptbl_h fmt =
   pp_header false fmt;
   Array.iter (fun (name, _) ->
       Format.fprintf fmt "  &&coq_lbl_%s,@." name
     ) opcodes
 
-let pp_coq_arity_h fmt =
+let pp_rocq_arity_h fmt =
   pp_header false fmt;
   Format.fprintf fmt "static signed char arity[] = {@.";
   Array.iter (fun (_, arity) ->
@@ -213,9 +213,9 @@ let usage () =
 
 let main () =
   match Sys.argv.(1) with
-  | "enum" -> pp_coq_instruct_h Format.std_formatter
-  | "jump" -> pp_coq_jumptbl_h Format.std_formatter
-  | "arity" -> pp_coq_arity_h Format.std_formatter
+  | "enum" -> pp_rocq_instruct_h Format.std_formatter
+  | "jump" -> pp_rocq_jumptbl_h Format.std_formatter
+  | "arity" -> pp_rocq_arity_h Format.std_formatter
   | "copml" -> pp_vmopcodes_ml Format.std_formatter
   | "copmli" -> pp_vmopcodes_mli Format.std_formatter
   | _ -> usage ()
